refactor(web): extract shared PluginDetail metadata/actions/state (Codex P1)
