+++ clients/web/src/domains/intelligence/components/plugins/plugin-detail.tsx
@@ -1,26 +1,20 @@
-import {
-    ArrowDownToLine,
-    ArrowLeft,
-    ArrowUpCircle,
-    Download,
-    ExternalLink,
-    Loader2,
-    Trash2,
-    TriangleAlert,
-} from "lucide-react";
-import { useState } from "react";
+import { ArrowLeft } from "lucide-react";
 
 import { FileMarkdown } from "@/components/file-markdown";
 import { PluginIcon } from "@/domains/intelligence/components/plugins/plugin-icon";
+import {
+    PluginDetailActionError,
+    PluginDetailActions,
+    PluginDetailError,
+    PluginDetailLoading,
+    PluginDetailMetadata,
+} from "@/domains/intelligence/components/plugins/plugin-detail-shared";
 import { PluginOriginBadge } from "@/domains/intelligence/components/plugins/plugin-origin-badge";
 import { UpdateAvailableBadge } from "@/domains/intelligence/components/plugins/update-available-badge";
-import {
-    shortSha,
-    usePluginDetail,
-} from "@/domains/intelligence/plugins/use-plugin-detail";
+import { usePluginDetail } from "@/domains/intelligence/plugins/use-plugin-detail";
 import type { PluginDrift } from "@/domains/intelligence/use-plugin-drift";
 import type { PluginsByNameGetResponse } from "@/generated/daemon/types.gen";
-import { Button, Card, ConfirmDialog } from "@vellumai/design-library";
+import { Button, Card } from "@vellumai/design-library";
 
 interface PluginDetailProps {
   assistantId: string;
@@ -38,13 +32,12 @@ interface PluginDetailProps {
  *
  * Renders the plugin's README plus the metadata we track (source, homepage,
  * license, contributed surfaces) and Install / Download / Upgrade / Remove
- * actions. Plugins have no file-list endpoint, so there is no file tree —
+ * actions. The metadata table, state views, and action set come from
+ * `plugin-detail-shared` so they stay in lockstep with the mobile detail.
+ * Plugins have no file-list endpoint, so there is no file tree —
  * README + metadata only.
  */
 export function PluginDetail({ assistantId, name, onBack }: PluginDetailProps) {
-  const [confirmingRemove, setConfirmingRemove] = useState(false);
-  const [confirmingUpgrade, setConfirmingUpgrade] = useState(false);
-
   const {
     plugin,
     drift,
@@ -62,26 +55,6 @@ export function PluginDetail({ assistantId, name, onBack }: PluginDetailProps) {
     hasLocalEdits,
   } = usePluginDetail(assistantId, name, { onRemoved: onBack });
 
-  const confirmRemove = () => {
-    setConfirmingRemove(false);
-    remove();
-  };
-
-  // Local edits would be clobbered by the re-install, so confirm first;
-  // a clean copy upgrades directly.
-  const handleUpgrade = () => {
-    if (hasLocalEdits) {
-      setConfirmingUpgrade(true);
-      return;
-    }
-    upgrade();
-  };
-
-  const confirmUpgrade = () => {
-    setConfirmingUpgrade(false);
-    upgrade();
-  };
-
   return (
     <div className="flex h-full min-h-0 flex-1 flex-col">
       <div className="mb-4 flex items-start gap-3">
@@ -97,16 +70,17 @@ export function PluginDetail({ assistantId, name, onBack }: PluginDetailProps) {
           plugin={plugin}
           drift={drift}
           onInstall={install}
-          onRemove={() => setConfirmingRemove(true)}
-          onUpgrade={handleUpgrade}
+          onRemove={remove}
+          onUpgrade={upgrade}
           isInstalling={isInstalling}
           isRemoving={isRemoving}
           isUpgrading={isUpgrading}
+          hasLocalEdits={hasLocalEdits}
         />
       </div>
 
       {(isInstallError || isRemoveError || isUpgradeError) && (
-        <ActionError
+        <PluginDetailActionError
           message={
             isInstallError
               ? "Failed to install plugin. Please try again."
@@ -120,12 +94,15 @@ export function PluginDetail({ assistantId, name, onBack }: PluginDetailProps) {
       <Card.Root asChild>
         <div className="min-h-0 flex-1 overflow-y-auto px-6 py-5">
           {isLoading ? (
-            <LoadingState />
+            <PluginDetailLoading />
           ) : isError || !plugin ? (
-            <DetailErrorState />
+            <PluginDetailError />
           ) : (
             <>
-              <Metadata plugin={plugin} surfaces={drift?.surfaces ?? null} />
+              <PluginDetailMetadata
+                plugin={plugin}
+                surfaces={drift?.surfaces ?? null}
+              />
               {plugin.readme ? (
                 <FileMarkdown content={plugin.readme} />
               ) : (
@@ -140,26 +117,6 @@ export function PluginDetail({ assistantId, name, onBack }: PluginDetailProps) {
           )}
         </div>
       </Card.Root>
-
-      <ConfirmDialog
-        open={confirmingRemove}
-        title="Remove plugin"
-        message={`Remove "${plugin?.name ?? name}" from this assistant?`}
-        confirmLabel="Remove"
-        destructive
-        onConfirm={confirmRemove}
-        onCancel={() => setConfirmingRemove(false)}
-      />
-
-      <ConfirmDialog
-        open={confirmingUpgrade}
-        title="Upgrade plugin"
-        message={`"${plugin?.name ?? name}" has local edits that will be overwritten by the upgrade. Continue?`}
-        confirmLabel="Upgrade anyway"
-        destructive
-        onConfirm={confirmUpgrade}
-        onCancel={() => setConfirmingUpgrade(false)}
-      />
     </div>
   );
 }
@@ -174,6 +131,7 @@ interface HeaderProps {
   isInstalling: boolean;
   isRemoving: boolean;
   isUpgrading: boolean;
+  hasLocalEdits: boolean;
 }
 
 function Header({
@@ -186,16 +144,10 @@ function Header({
   isInstalling,
   isRemoving,
   isUpgrading,
+  hasLocalEdits,
 }: HeaderProps) {
-  const installed = plugin?.installed ?? false;
   const isExternal = plugin?.source?.kind === "github";
-  const artifact = plugin?.artifact ?? null;
   const updateAvailable = drift?.status === "update-available";
-  const upgradeTitle = updateAvailable
-    ? `Upgrade ${shortSha(drift?.local?.commit ?? null)} → ${shortSha(
-        drift?.remote?.commit ?? null,
-      )}`
-    : undefined;
 
   return (
     <div className="flex min-w-0 flex-1 flex-col gap-3 sm:flex-row sm:items-center">
@@ -232,192 +184,18 @@ function Header({
       </div>
 
       {plugin ? (
-        installed ? (
-          <div className="flex shrink-0 items-center gap-2">
-            {artifact ? (
-              <Button asChild leftIcon={<Download aria-hidden />}>
-                <a href={artifact.url} download>
-                  {artifact.label ?? "Download"}
-                </a>
-              </Button>
-            ) : null}
-            {updateAvailable ? (
-              <Button
-                type="button"
-                onClick={onUpgrade}
-                disabled={isUpgrading}
-                title={upgradeTitle}
-                leftIcon={
-                  isUpgrading ? (
-                    <Loader2 className="animate-spin" aria-hidden />
-                  ) : (
-                    <ArrowUpCircle aria-hidden />
-                  )
-                }
-              >
-                Upgrade
-              </Button>
-            ) : null}
-            <Button
-              type="button"
-              variant="dangerOutline"
-              onClick={onRemove}
-              disabled={isRemoving}
-              leftIcon={
-                isRemoving ? (
-                  <Loader2 className="animate-spin" aria-hidden />
-                ) : (
-                  <Trash2 aria-hidden />
-                )
-              }
-            >
-              Remove
-            </Button>
-          </div>
-        ) : (
-          <Button
-            type="button"
-            onClick={onInstall}
-            disabled={isInstalling}
-            leftIcon={
-              isInstalling ? (
-                <Loader2 className="animate-spin" aria-hidden />
-              ) : (
-                <ArrowDownToLine aria-hidden />
-              )
-            }
-          >
-            Install
-          </Button>
-        )
+        <PluginDetailActions
+          plugin={plugin}
+          drift={drift}
+          onInstall={onInstall}
+          onRemove={onRemove}
+          onUpgrade={onUpgrade}
+          isInstalling={isInstalling}
+          isRemoving={isRemoving}
+          isUpgrading={isUpgrading}
+          hasLocalEdits={hasLocalEdits}
+        />
       ) : null}
-    </div>
-  );
-}
-
-function Metadata({
-  plugin,
-  surfaces,
-}: {
-  plugin: PluginsByNameGetResponse;
-  surfaces: PluginDrift["surfaces"];
-}) {
-  const repo = plugin.source?.kind === "github" ? plugin.source.repo : "Local";
-  const repoHref =
-    plugin.source?.kind === "github"
-      ? `https://github.com/${plugin.source.repo}`
-      : null;
-
-  const rows: { label: string; value: string; href?: string }[] = [
-    {
-      label: "Source",
-      value: repo,
-      href: repoHref ?? undefined,
-    },
-  ];
-  if (plugin.homepage) {
-    rows.push({
-      label: "Homepage",
-      value: plugin.homepage,
-      href: plugin.homepage,
-    });
-  }
-  if (plugin.license) {
-    rows.push({ label: "License", value: plugin.license });
-  }
-  // Surfaces are only present for an installed copy; list the non-empty
-  // contributions (skills / hooks / tools) so the panel shows what the
-  // plugin actually adds.
-  if (surfaces) {
-    if (surfaces.skills.length > 0) {
-      rows.push({ label: "Skills", value: String(surfaces.skills.length) });
-    }
-    if (surfaces.hooks.length > 0) {
-      rows.push({ label: "Hooks", value: String(surfaces.hooks.length) });
-    }
-    if (surfaces.tools.length > 0) {
-      rows.push({ label: "Tools", value: String(surfaces.tools.length) });
-    }
-  }
-
-  return (
-    <dl
-      className="mb-5 grid gap-x-6 gap-y-2 border-b pb-5 sm:grid-cols-[max-content_1fr]"
-      style={{ borderColor: "var(--border-base)" }}
-    >
-      {rows.map((row) => (
-        <div key={row.label} className="contents">
-          <dt
-            className="text-body-small-default"
-            style={{ color: "var(--content-tertiary)" }}
-          >
-            {row.label}
-          </dt>
-          <dd
-            className="min-w-0 truncate text-body-small-default"
-            style={{ color: "var(--content-secondary)" }}
-          >
-            {row.href ? (
-              <a
-                href={row.href}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-flex items-center gap-1 underline"
-                style={{ color: "var(--primary-base, #60a5fa)" }}
-              >
-                {row.value}
-                <ExternalLink className="h-3 w-3" aria-hidden />
-              </a>
-            ) : (
-              row.value
-            )}
-          </dd>
-        </div>
-      ))}
-    </dl>
-  );
-}
-
-function LoadingState() {
-  return (
-    <div className="flex items-center justify-center py-12">
-      <Loader2
-        className="h-6 w-6 animate-spin"
-        style={{ color: "var(--content-tertiary)" }}
-      />
-    </div>
-  );
-}
-
-function DetailErrorState() {
-  return (
-    <div
-      className="flex flex-col items-center justify-center gap-2 py-12 text-center"
-      style={{ color: "var(--content-tertiary)" }}
-    >
-      <TriangleAlert className="h-6 w-6" aria-hidden />
-      <p className="text-body-medium-default">
-        We couldn&apos;t load this plugin.
-      </p>
-      <p className="text-body-small-default">
-        It may not exist, or your assistant may be on an older build.
-      </p>
-    </div>
-  );
-}
-
-function ActionError({ message }: { message: string }) {
-  return (
-    <div
-      className="mb-3 flex items-center gap-2 rounded px-3 py-2 text-body-small-default"
-      style={{
-        backgroundColor: "var(--surface-secondary)",
-        color: "var(--content-warning, var(--content-tertiary))",
-      }}
-      role="alert"
-    >
-      <TriangleAlert className="h-4 w-4 shrink-0" aria-hidden />
-      {message}
     </div>
   );
 }
